+++ rest/src/decorators/index.ts
@@ -1,3 +1,6 @@
+export * from './errorPassThrough.decorator';
+export * from './handle400WithValidations.decorator';
 export * from './handle404.decorator';
 export * from './ignoreClientErrors.decorator';
-export * from './handle400WithValidations.decorator';
+export * from './suppressError.decorator';
+export * from './withRestClientError.decorator';