+++ rest/src/index.ts
@@ -1,7 +1,9 @@
 export * from './decorators';
 export * from './middlewares';
 
+export * from './misc/clientErrorHandlingMiddleware.options';
 export * from './misc/defaults';
 export * from './misc/middlewareTypes';
 export * from './misc/restHttpError.interface';
+export * from './misc/tokens';
 export * from './misc/withRestClientContext';